# docs/openapi/spec.yaml
openapi: 3.1.0
info:
  title: Eazy Recycling Backend Service
  version: 0.0.1
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
paths:
...
  /products/{id}:
    get:
      tags:
      - product-controller
      operationId: getProductById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductResponse"
    put:
      tags:
      - product-controller
      operationId: updateProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductResponse"
    delete:
      tags:
      - product-controller
      operationId: deleteProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
  /product-categories/{id}:
    get:
      tags:
      - product-category-controller
      operationId: getCategoryById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductCategoryResponse"
    put:
      tags:
      - product-category-controller
      operationId: updateCategory
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCategoryRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductCategoryResponse"
    delete:
      tags:
      - product-category-controller
      operationId: deleteCategory
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
...
  /products:
    get:
      tags:
      - product-controller
      operationId: getAllProducts
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductResponse"
    post:
      tags:
      - product-controller
      operationId: createProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductResponse"
  /product-categories:
    get:
      tags:
      - product-category-controller
      operationId: getAllCategories
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductCategoryResponse"
    post:
      tags:
      - product-category-controller
      operationId: createCategory
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCategoryRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductCategoryResponse"
...
components:
  schemas:
...
    ProductRequest:
      type: object
      properties:
        code:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        categoryId:
          type: string
          format: uuid
        unitOfMeasure:
          type: string
          minLength: 1
        vatCode:
          type: string
          minLength: 1
        salesAccountNumber:
          type: string
        purchaseAccountNumber:
          type: string
        status:
          type: string
          minLength: 1
        defaultPrice:
          type: number
      required:
      - code
      - name
      - status
      - unitOfMeasure
      - vatCode
    ProductResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        categoryId:
          type: string
          format: uuid
        unitOfMeasure:
          type: string
        vatCode:
          type: string
        salesAccountNumber:
          type: string
        purchaseAccountNumber:
          type: string
        status:
          type: string
        defaultPrice:
          type: number
        createdAt:
          type: string
        createdByName:
          type: string
        updatedAt:
          type: string
        updatedByName:
          type: string
      required:
      - code
      - id
      - name
      - status
      - unitOfMeasure
      - vatCode
    ProductCategoryRequest:
      type: object
      properties:
        code:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        description:
          type: string
      required:
      - code
      - name
    ProductCategoryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
        createdByName:
          type: string
        updatedAt:
          type: string
        updatedByName:
          type: string
      required:
      - code
      - id
      - name
...
  securitySchemes:
    bearerAuth:
      type: http
      name: Authorization
      in: header
      scheme: bearer
      bearerFormat: JWT